backend/lambdas/auth/requirements.txt
bcrypt==4.1.1
PyJWT==2.12.1
